+++ teacherPal/SchoolBusPhoto/SchoolBusPhoto/GKAlreadyCell.m
@@ -99,7 +99,7 @@
     }];
     
 }
--(NSString *)stateStr:(int )a
+-(NSString *)stateStr:(NSInteger )a
 {
     switch (a) {
         case 1:
