+++ teacherPal/SchoolBusPhoto/SchoolBusPhoto/GKPhotoTagScrollView.m
@@ -162,7 +162,7 @@
         [btn setTitleColor:[UIColor blackColor] forState:UIControlStateNormal];
         
         for (int j=0; j<[arr count]; j++) {
-            int tag=[[arr objectAtIndex:j] integerValue];
+            long tag=[[arr objectAtIndex:j] integerValue];
             if(tag==[btn.phototag.tagid integerValue])
             {
 //                [btn setBackgroundImage:[[UIImage imageNamed:@"navbar-button-blue-active"] stretchableImageWithLeftCapWidth:3 topCapHeight:15] forState:UIControlStateNormal];
