+++ teacherPal/SchoolBusPhoto/SchoolBusPhoto/MTCustomActionSheet.h
@@ -29,7 +29,7 @@
 {
     UIView *mainView;
     UIButton *cancelBtn;
-    int count;
+    NSInteger count;
     UIDatePicker *datepicker;
 }
