+++ teacherPal/SchoolBusPhoto/SchoolBusPhoto/NSDate+convenience.h
@@ -12,7 +12,7 @@
 -(NSDate *)offsetMonth:(int)numMonths;
 -(NSDate *)offsetDay:(int)numDays;
 -(NSDate *)offsetHours:(int)hours;
--(int)numDaysInMonth;
+-(NSUInteger)numDaysInMonth;
 -(int)firstWeekDayInMonth;
 -(int)year;
 -(int)month;
